requirements.txt
py4j==0.10.9.9
pyspark==4.0.0
cassandra-driver~=3.29.2
redis~=6.2.0
fastapi[standard]~=0.116.1
uvicorn[standard]~=0.35.0
requests~=2.32.4